nästan klar

+++ src/main/java/Operations.java
@@ -1,7 +1,7 @@
 public class Operations {
 
     public static double divide(int numerator, int denominator){
-        double result = numerator / denominator;
+        float result = numerator / denominator;
         return result;
     };
